js 파일 종류별로 더 만듬
